% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_palettes}
\alias{country_palettes}
\title{Color palettes of country flags.}
\format{
A data frame with 195 rows and 4 variables:
\describe{
  \item{country}{name of country in English}
  \item{iso2}{ISO 3166-1 Alpha-2 code of country}
  \item{iso3}{ISO 3166-1 Alpha-3 code of country}
  \item{palette}{vector of hex color codes of the country flag}
}
}
\source{
Adapted from \url{https://vexillology.fandom.com/wiki/User:HansLN/National_flag_colors}
}
\usage{
country_palettes
}
\description{
A dataset containing the colors of the flags of 195 countries.
}
\keyword{datasets}
