% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_palette.R
\name{expand_palette}
\alias{expand_palette}
\title{Expand palette}
\usage{
expand_palette(palette)
}
\arguments{
\item{palette}{A color palette (vector of hex colors) to be expanded}
}
\value{
An expanded color palette
}
\description{
This function inputs a color palette (vector) and expands the colors by repeating them in lighter shades
}
