% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_image.R
\name{flag_image}
\alias{flag_image}
\title{Make a flag image}
\usage{
flag_image(image, country = "SE", destination = "", scale = 100, bg = "black")
}
\arguments{
\item{image}{An image filename or url}

\item{country}{The name of a country, either official English name (e.g. "Sweden"), or ISO-2/3 (e.g. "SE"/"SWE")}

\item{destination}{Optional path/filename for saving output image}

\item{scale}{Scale the image (default=100, i.e. 100pt)}

\item{bg}{Set background if alpha (default="black"; options are "white", "none")}
}
\value{
A data frame
}
\description{
This function inputs an image (file or url) and outputs a new
image with coloring matching a country flag
}
