% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_custom_palette.R
\name{get_custom_palette}
\alias{get_custom_palette}
\title{Get custom palette}
\usage{
get_custom_palette(country = "Sweden", n = 3)
}
\arguments{
\item{country}{The name of a country, either official English name (e.g. "Sweden"), or ISO-2/3 (e.g. "SE"/"SWE")}

\item{n}{The number of colors required}
}
\value{
A color palette
}
\description{
This function inputs a country name and a required palette length
and returns a color palette based on the country's flag with an
expanded set of colors (if needed) and colors sorted
}
