% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{get_palette}
\alias{get_palette}
\title{Get palette}
\usage{
get_palette(country)
}
\arguments{
\item{country}{The name of a country, either official English name (e.g. "Sweden"), or ISO-2/3 (e.g. "SE"/"SWE")}
}
\value{
A color palette
}
\description{
This function inputs a country name returns a color palette based on the country's flag
}
