% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_image_df.R
\name{make_image_df}
\alias{make_image_df}
\title{Make image data frame}
\usage{
make_image_df(image, scale = 100, bg = "black")
}
\arguments{
\item{image}{An image filename or url}

\item{scale}{Scale the image (default=100, i.e. 100pt)}

\item{bg}{Set background if alpha (default="black"; options are "white", "none")}
}
\value{
A data frame
}
\description{
This function inputs an image (file or url) and outputs a data frame
with coloring data for plotting
}
